requirements.txt
requests
streamlit-aggrid
pandas
numpy
plotly